+++ qatrack/notifications/admin/__init__.py
@@ -1,4 +1,5 @@
 import qatrack.notifications.common.admin
+import qatrack.notifications.parts.admin
 import qatrack.notifications.qccompleted.admin
 import qatrack.notifications.qcreview.admin
 import qatrack.notifications.qcscheduling.admin